# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Task Management System API
  description: |-
    This is a Task Management System API built with Java and Spring.
    The system provides functionality for creating, editing, deleting, and viewing tasks with authentication and authorization via JWT tokens.
    
    Features:
    - User authentication and authorization
    - Role-based access control (Admin and User roles)
    - Task management with status tracking, priorities, and comments
    - Filtering and pagination for task retrieval
    - Comprehensive error handling and input validation
    
    Roles:
    - Admin: Full access to all tasks and user management
    - User: Access to own tasks and tasks where assigned as executor
    
    Authentication: All endpoints (except login/register) require JWT token in Authorization header.
  version: 1.0.0
  contact:
    email: support@taskmanagement.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
...
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.taskmanagement.com/v1
    description: Production server
paths:
...
  /task/tasks:
    get:
      tags:
        - Tasks
      summary: Get paginated list of tasks
      description: Retrieve all tasks with pagination. Accessible by USER and ADMIN roles.
      security:
        - bearerAuth: [ ]
      parameters:
        - name: page
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            default: 3
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageTaskResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                unauthorized:
                  $ref: '#/components/examples/ErrorResponseUnauthorizedExample'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /task/{id}:
    get:
      tags:
        - Tasks
      summary: Get task by ID
      description: Retrieve a specific task by its ID. Accessible by USER and ADMIN roles.
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Task not found
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                unauthorized:
                  $ref: '#/components/examples/ErrorResponseUnauthorizedExample'
        '403':
          description: Forbidden
...
  /task/:
    post:
      tags:
        - Tasks
      summary: Create new task
      description: Create a new task. Accessible only by ADMIN role.
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Invalid input data
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                unauthorized:
                  $ref: '#/components/examples/ErrorResponseUnauthorizedExample'
        '403':
          description: Forbidden

    put:
      tags:
        - Tasks
      summary: Update task
      description: Update an existing task. Accessible only by ADMIN role.
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Invalid input data
        '404':
          description: Task not found
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                unauthorized:
                  $ref: '#/components/examples/ErrorResponseUnauthorizedExample'
        '403':
          description: Forbidden

    delete:
      tags:
        - Tasks
      summary: Delete task
      description: Delete a task. Accessible only by ADMIN role.
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '204':
          description: Task deleted successfully
        '404':
          description: Task not found
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                unauthorized:
                  $ref: '#/components/examples/ErrorResponseUnauthorizedExample'
        '403':
          description: Forbidden
  /task/{id}/status:
    patch:
      tags:
        - Tasks
      summary: Update task status
      description: |
        Update the status of a task status. Accessible by USER and ADMIN roles.
        - Users can update task status only for their own tasks (where they are assigned as executor)
        - Admins can update task status for any task
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          example: 11
        - name: statusId
          in: query
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          example: 1
      responses:
        '200':
          description: Status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Task or status not found
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                unauthorized:
                  $ref: '#/components/examples/ErrorResponseUnauthorizedExample'
        '403':
          description: Forbidden
...
  /status/statuses:
    get:
      tags:
        - Task Statuses
      summary: Get all task statuses
      description: Retrieve all task statuses with sorting options
      parameters:
        - name: sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            enum: [ ASC, DESC ]
            default: ASC
        - name: sortBy
          in: query
          description: Field to sort by (id, name, updatedAt)
          required: false
          schema:
            type: string
            default: id
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskStatusResponseDto'
        '400':
          description: Invalid sort parameters
  /status/{id}:
    get:
      tags:
        - Task Statuses
      summary: Get task status by ID
      description: Retrieve a specific task status by its ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatusResponseDto'
        '404':
          description: Task status not found
    put:
      tags:
        - Task Statuses
      summary: Update task status
      description: Update an existing task status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskStatusRequestDto'
      responses:
        '200':
          description: Task status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatusResponseDto'
        '400':
          description: Invalid input data
        '404':
          description: Task status not found
    delete:
      tags:
        - Task Statuses
      summary: Delete task status
      description: Delete a task status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Task status deleted successfully
          content:
            application/json:
              schema:
                type: string
                example: "Task status deleted"
        '404':
          description: Task status not found
  /status:
    post:
      tags:
        - Task Statuses
      summary: Create new task status
      description: Create a new task status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskStatusRequestDto'
      responses:
        '201':
          description: Task status created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatusResponseDto'
        '400':
          description: Invalid input data

  # users
  /user/users:
    get:
      tags:
        - Users
      summary: Get all users
      description: Retrieve all users with sorting options
      parameters:
        - name: sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            enum: [ ASC, DESC ]
            default: ASC
        - name: sortBy
          in: query
          description: Field to sort by (id, email, name, createdAt, updatedAt)
          required: false
          schema:
            type: string
            default: id
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponseDto'
        '400':
          description: Invalid sort parameters
  /user/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: Retrieve a specific user by its ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '404':
          description: User not found
    put:
      tags:
        - Users
      summary: Update user
      description: Update an existing user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequestDto'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '400':
          description: Invalid input data
        '404':
          description: User not found
    delete:
      tags:
        - Users
      summary: Delete user
      description: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: string
                example: "User deleted"
        '404':
          description: User not found
  /user/by-email/{email}:
    get:
      tags:
        - Users
      summary: Get user by email
      description: Retrieve a specific user by email address
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            format: email
            example: "user@example.com"
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '404':
          description: User not found
  /user:
    post:
      tags:
        - Users
      summary: Create new user
      description: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequestDto'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '400':
          description: Invalid input data
        '409':
          description: User with this email already exists
components:
  schemas:
...
    TaskRequest:
      type: object
      required:
        - title
        - description
        - taskStatusId
        - taskPriorityId
        - authorId
      properties:
        title:
          type: string
          maxLength: 50
          example: "Implement user authentication"
          description: Task title
        description:
          type: string
          maxLength: 350
          example: "Implement JWT-based authentication system with Spring Security"
          description: Task description
        taskStatusId:
          type: integer
          format: int64
          example: 1
          description: ID of the task status
        taskPriorityId:
          type: integer
          format: int64
          example: 2
          description: ID of the task priority
        authorId:
          type: integer
          format: int64
          example: 1
          description: ID of the task author
        assigneeId:
          type: integer
          format: int64
          nullable: true
          example: 2
          description: ID of the task assignee (optional)
    TaskResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
          description: Task ID
        title:
          type: string
          example: "Implement user authentication"
          description: Task title
        description:
          type: string
          example: "Implement JWT-based authentication system with Spring Security"
          description: Task description
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15 10:30:00"
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15 11:45:00"
          description: Last update timestamp
        status:
          type: string
          example: "IN_PROGRESS"
          description: Task status name
        priority:
          type: string
          example: "HIGH"
          description: Task priority name
        author:
          type: string
          example: "john_doe"
          description: Author username
        assignee:
          type: string
          nullable: true
          example: "jane_smith"
          description: Assignee username
        comments:
          type: array
          items:
            $ref: '#/components/schemas/TaskCommentResponse'
          description: List of task comments
    TaskCommentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
          description: Comment ID
        content:
          type: string
          example: "This task is progressing well"
          description: Comment content
        author:
          type: string
          example: "john_doe"
          description: Comment author username
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T12:00:00Z"
          description: Comment creation timestamp
    PageTaskResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/TaskResponse'
          description: Page content
        totalElements:
          type: integer
          format: int64
          example: 25
          description: Total number of elements
        totalPages:
          type: integer
          format: int32
          example: 9
          description: Total number of pages
        size:
          type: integer
          format: int32
          example: 3
          description: Number of elements per page
        number:
          type: integer
          format: int32
          example: 0
          description: Current page number (0-based)
        first:
          type: boolean
          example: true
          description: Is first page
        last:
          type: boolean
          example: false
          description: Is last page
        numberOfElements:
          type: integer
          format: int32
          example: 3
          description: Number of elements in current page
        empty:
          type: boolean
          example: false
          description: Is page empty
    RoleResponse:
      type: object
      description: Role entity representing user roles in the system.
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the role (auto-generated).
          example: 1
          readOnly: true
        name:
          type: string
          description: Name of the role (e.g., ADMIN, USER).
          example: "USER"
      required:
        - name

    # errors
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15 10:30:00"
          description: Error timestamp
        status:
          type: integer
          example: 400
          description: HTTP status code
        error:
          type: string
          example: "Bad Request"
          description: Error type
        message:
          type: string
          example: "Validation failed"
          description: Error message
        path:
          type: string
          example: "/auth/sign-up"
          description: Request path
        errors:
          type: array
          items:
            type: string
          example: [ "Password must be at least 8 characters long" ]
          description: Detailed error messages
...
    TaskStatusRequestDto:
      type: object
      required:
        - name
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          example: 1
          description: Task status ID (optional for create, required for update)
        name:
          type: string
          maxLength: 30
          example: "In Progress"
          description: Status name
    TaskStatusResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
          description: Task status ID
        name:
          type: string
          example: "In Progress"
          description: Status name
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15 10:30:00"
          description: Last update timestamp

    # users
    UserRequestDto:
      type: object
      required:
        - email
        - password
        - name
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          example: 1
          description: User ID (optional for create, required for update)
        email:
          type: string
          format: email
          maxLength: 50
          example: "user@example.com"
          description: User email address
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 255
          example: "password123"
          description: User password
        name:
          type: string
          maxLength: 30
          example: "John Doe"
          description: User full name
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleResponse'
          description: User roles
    UserResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
          description: User ID
        email:
          type: string
          example: "user@example.com"
          description: User email address
        name:
          type: string
          example: "John Doe"
          description: User full name
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15 10:30:00"
          description: User creation timestamp
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15 10:30:00"
          description: Last update timestamp
        authoredTaskIds:
          type: array
          items:
            type: integer
            format: int64
          example: [ 1, 2, 3 ]
          description: IDs of tasks authored by this user
        assignedTaskIds:
          type: array
          items:
            type: integer
            format: int64
          example: [ 4, 5, 6 ]
          description: IDs of tasks assigned to this user
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleResponse'
          description: User roles

  # examples
  examples:
    ErrorResponseUnauthorizedExample:
      summary: Example of general unauthorized
      value:
        type: "/errors/unauthorized"
        title: "unauthorized"
        status: 401
        detail: "Access is denied. Authorization is required."
        instance: "/api/v1/auth/sign-in"
        timestamp: "2025-09-04 19:10:01"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: [ ]
...